refactor: Organize files and separate the server operations from action to more readable code
